backend/requirements.txt
# ============================================
# FinLex AI - Backend Dependencies
# ============================================

# --- Web Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# --- Database ---
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.14.1

# --- Redis ---
redis==5.2.1

# --- AI / LLM ---
langchain==1.2.15
langchain-classic==1.0.3
langchain-core==1.2.28
langchain-openai==1.1.12
langchain-google-genai==4.2.1
langchain-community==0.4.1
langchain-text-splitters==1.1.1
openai==2.31.0
google-genai==1.72.0

# --- Vector Database ---
langchain-chroma==1.1.0
chromadb==1.5.7

# --- Document Processing ---
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5

# --- Auth ---
pyjwt[crypto]==2.10.1
httpx==0.28.1

# --- Payments ---
razorpay==1.4.2

# --- Validation & Serialization ---
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# --- Security ---
passlib[bcrypt]==1.7.4

# --- Utilities ---
python-dotenv==1.0.1
tenacity==9.0.0
aiofiles==24.1.0

# --- Testing ---
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# --- Development ---
ruff==0.8.6
